Override unexpected changes to wrapped light
This commit adds a timeout to calls it makes to the wrapped light
entity. If an update is received on the light entity outside of this
timer it is marked as an 'unexpected' update and the sequences will
be reset.

This should restore the light to the correct state after, eg, a power
outage.